+++ src/modules/rlm_cache/drivers/rlm_cache_rbtree/rlm_cache_rbtree.c
@@ -35,7 +35,7 @@ typedef struct rlm_cache_rbtree {
 
 typedef struct rlm_cache_rbtree_entry {
 	rlm_cache_entry_t	fields;		//!< Entry data.
-	size_t			heap_id;	//!< Offset used for heap.
+	int32_t			heap_id;	//!< Offset used for heap.
 } rlm_cache_rbtree_entry_t;
 
 /** Compare two entries by key
